- Jorrit fixed awstest to compile again after changing the Broadcast 	  prototype.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22929 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/iaws/awsecomp.h
@@ -64,7 +64,7 @@ public:
   { return comp->UnregisterSlot (slot, signal); }
 
   /// Broadcasts a signal to all slots that are interested.
-  virtual void Broadcast (unsigned long signal)
+  virtual void Broadcast (uint32 signal)
   { comp->Broadcast (signal); }
 
 public:
